# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing Cloud Hypervisor VMs.",
        "title": "Cloud Hypervisor API",
        "termsOfService": "https://github.com/org/ch-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/org/ch-api/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms/{id}/console": {
            "get": {
                "description": "Streams the serial console output of the virtual machine over WebSocket.",
                "tags": [
                    "Console"
                ],
                "summary": "VM Console",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "WebSocket upgrade"
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vms/{id}/disks": {
            "post": {
                "description": "Attaches a new disk to the virtual machine with the given ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disks"
                ],
                "summary": "Add Disk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Disk configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddDiskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.DiskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "404": {
                        "description": "VM or disk not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/disks/{disk_id}": {
            "delete": {
                "description": "Detaches the disk from the virtual machine with the given ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disks"
                ],
                "summary": "Remove Disk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Disk ID",
                        "name": "disk_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.DiskResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "VM or disk not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/disks/{disk_id}/snapshot": {
            "post": {
                "description": "Creates a snapshot of the disk attached to the virtual machine.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disks"
                ],
                "summary": "Snapshot Disk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Disk ID",
                        "name": "disk_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.DiskResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "VM or disk not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.AddDiskRequest": {
            "type": "object",
            "properties": {
                "direct": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.DiskResponse": {
            "type": "object",
            "properties": {
                "direct": {
                    "type": "boolean"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                }
            }
        },
...
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "errors": {
                    "description": "Errors holds machine-readable validation errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/problem.FieldError"
                    }
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code generated by the origin server.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        },
        "problem.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}