# pkg/api/problem/problem.go
package problem

import (
	"encoding/json"
	"fmt"
	"net/http"
)

// Detail is an RFC 7807 Problem Details JSON object.
// See https://tools.ietf.org/html/rfc7807
type Detail struct {
	// Type is a URI reference that identifies the problem type.
	Type string `json:"type"`

	// Title is a short, human-readable summary of the problem type.
	Title string `json:"title"`

	// Status is the HTTP status code generated by the origin server.
	Status int `json:"status"`

	// Detail is a human-readable explanation specific to this occurrence.
	Detail string `json:"detail,omitempty"`

	// Instance is a URI reference that identifies the specific occurrence.
	Instance string `json:"instance,omitempty"`

	// Errors holds machine-readable validation errors.
	Errors []FieldError `json:"errors,omitempty"`
}

// FieldError describes a single validation failure.
type FieldError struct {
	Field   string `json:"field"`
	Message string `json:"message"`
}

// Write encodes p as JSON and writes it to w with the given status code.
// The Content-Type header is set to application/problem+json.
func (p *Detail) Write(w http.ResponseWriter) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(p.Status)
	_ = json.NewEncoder(w).Encode(p)
}

// ---------------------------------------------------------------------------
// Constructors
// ---------------------------------------------------------------------------

const baseType = "https://github.com/org/ch-api/errors"

// BadRequest returns a 400 Bad Request problem detail.
func BadRequest(instance, detail string, errs ...FieldError) *Detail {
	return &Detail{
		Type:     baseType + "/invalid-request",
		Title:    "Invalid Request",
		Status:   http.StatusBadRequest,
		Detail:   detail,
		Instance: instance,
		Errors:   errs,
	}
}

// NotFound returns a 404 Not Found problem detail.
func NotFound(instance, detail string) *Detail {
	return &Detail{
		Type:     baseType + "/not-found",
		Title:    "Not Found",
		Status:   http.StatusNotFound,
		Detail:   detail,
		Instance: instance,
	}
}

// InternalServerError returns a 500 Internal Server Error problem detail.
func InternalServerError(instance string) *Detail {
	return &Detail{
		Type:     baseType + "/internal-server-error",
		Title:    "Internal Server Error",
		Status:   http.StatusInternalServerError,
		Detail:   "An unexpected error occurred. Please try again later.",
		Instance: instance,
	}
}

// Unauthorized returns a 401 Unauthorized problem detail.
func Unauthorized(instance, detail string) *Detail {
	return &Detail{
		Type:     baseType + "/unauthorized",
		Title:    "Unauthorized",
		Status:   http.StatusUnauthorized,
		Detail:   detail,
		Instance: instance,
	}
}

// Forbidden returns a 403 Forbidden problem detail.
func Forbidden(instance, detail string) *Detail {
	return &Detail{
		Type:     baseType + "/forbidden",
		Title:    "Forbidden",
		Status:   http.StatusForbidden,
		Detail:   detail,
		Instance: instance,
	}
}

// Conflict returns a 409 Conflict problem detail.
func Conflict(instance, detail string) *Detail {
	return &Detail{
		Type:     baseType + "/conflict",
		Title:    "Conflict",
		Status:   http.StatusConflict,
		Detail:   detail,
		Instance: instance,
	}
}

// RateLimitExceeded returns a 429 Too Many Requests problem detail.
func RateLimitExceeded(instance, detail string, retryAfter int) *Detail {
	return &Detail{
		Type:     baseType + "/rate-limit-exceeded",
		Title:    "Too Many Requests",
		Status:   http.StatusTooManyRequests,
		Detail:   detail,
		Instance: instance,
	}
}

// UnprocessableEntity returns a 422 Unprocessable Entity problem detail.
func UnprocessableEntity(instance, detail string, errs ...FieldError) *Detail {
	return &Detail{
		Type:     baseType + "/unprocessable-entity",
		Title:    "Unprocessable Entity",
		Status:   http.StatusUnprocessableEntity,
		Detail:   detail,
		Instance: instance,
		Errors:   errs,
	}
}

// Field creates a single FieldError.
func Field(field, message string, args ...any) FieldError {
	return FieldError{
		Field:   field,
		Message: fmt.Sprintf(message, args...),
	}
}